This fixes 5 potential bugs where the conditional would always be true.
git-svn-id: https://svn.musicpd.org/mpd/trunk@4659 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/inputPlugins/aac_plugin.c
@@ -239,7 +239,7 @@ static float getAacFloatTotalTime(char *file)
 	unsigned int sampleRate;
 	unsigned char channels;
 	InputStream inStream;
-	size_t bread;
+	long bread;
 
 	if (openInputStream(&inStream, file) < 0)
 		return -1;
@@ -291,7 +291,7 @@ static int aac_decode(OutputBuffer * cb, DecoderControl * dc, char *path)
 	faacDecHandle decoder;
 	faacDecFrameInfo frameInfo;
 	faacDecConfigurationPtr config;
-	size_t bread;
+	long bread;
 	unsigned int sampleRate;
 	unsigned char channels;
 	int eof = 0;
